backend/requirements.txt
Flask
psycopg2-binary
flask-cors
numpy
scipy